requirements.txt
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
Flask==3.0.3
Flask-RESTful==0.3.10
greenlet==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.2
requests==2.28.1
six==1.16.0
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
